# frontend/public/docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FindEZ Integration API",
    "version": "1",
    "description": "Public integration API for inventory in linked Team Spaces. Documentation: https://findez.ai/docs/api. Examples use synthetic IDs and do not contain usable credentials.",
    "contact": {
      "name": "FindEZ support",
      "email": "info@findez.ai"
    }
  },
  "paths": {
    "/api/v1/keys": {
      "get": {
        "tags": [
          "Key management"
        ],
        "summary": "List key metadata",
        "operationId": "list_keys_api_v1_keys_get",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KeyList"
                },
                "example": {
                  "keys": [
                    {
                      "id": "40000000-0000-0000-0000-000000000001",
                      "workspace_id": "10000000-0000-0000-0000-000000000001",
                      "name": "Team reporting",
                      "key_prefix": "findez_live_sk_EXAMPL",
                      "scopes": [
                        "items:read",
                        "workspace:read"
                      ],
                      "created_at": "2026-09-10T12:00:00Z",
                      "last_used_at": null,
                      "expires_at": null,
                      "revoked_at": null
                    }
                  ]
                }
              }
            },
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "example": "no-store"
              }
            }
          },
          "default": {
            "description": "Error; see the error-code reference. User-session and gateway failures may have a string detail or non-JSON body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "UserSession": []
          }
        ],
        "description": "Requires a Supabase user access token. Returns that user's keys newest first, including expired and revoked keys. Neither the raw key nor its hash is returned. No pagination or status filter is supported. Cache-Control is no-store. last_used_at is approximate: authentication schedules an update at most once per five minutes, and a successful authentication may precede a later permission or request failure. A null timestamp means no recorded use, not proof of no requests. It is not an audit log or successful-write receipt.",
        "x-required-scope": null,
        "x-rate-bucket": null,
        "x-path-values": {},
        "x-query-example": {},
        "x-extra-examples": {}
      },
      "post": {
        "tags": [
          "Key management"
        ],
        "summary": "Create an API key",
        "operationId": "create_key_api_v1_keys_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateKeyRequest"
              },
              "example": {
                "name": "Team reporting",
                "workspace_id": "10000000-0000-0000-0000-000000000001",
                "scopes": [
                  "items:read",
                  "workspace:read"
                ],
                "expires_at": null
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssuedKey"
                },
                "example": {
                  "id": "40000000-0000-0000-0000-000000000001",
                  "workspace_id": "10000000-0000-0000-0000-000000000001",
                  "name": "Team reporting",
                  "key_prefix": "findez_live_sk_EXAMPL",
                  "scopes": [
                    "items:read",
                    "workspace:read"
                  ],
                  "created_at": "2026-09-10T12:00:00Z",
                  "last_used_at": null,
                  "expires_at": null,
                  "revoked_at": null,
                  "org_id": "20000000-0000-0000-0000-000000000001",
                  "created_by": "20000000-0000-0000-0000-000000000001",
                  "key": "findez_live_sk_EXAMPLE_NOT_A_REAL_KEY"
                }
              }
            },
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "example": "no-store"
              }
            }
          },
          "default": {
            "description": "Error; see the error-code reference. User-session and gateway failures may have a string detail or non-JSON body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "UserSession": []
          }
        ],
        "description": "Requires an owner's Supabase user access token, not an API key. With workspace_id, only workspace scopes are accepted and the caller must own that team. With workspace_id omitted or null, only org scopes are accepted and the caller must own at least one team. Scope names are trimmed and deduplicated. expires_at must be in the future; omitted/null means no expiration. Prefer an ISO 8601 timestamp with Z or an offset; a timezone-free value is interpreted as UTC. The web form defaults to 90 days, while this HTTP endpoint defaults to no expiration. Returns metadata plus the complete key once, with Cache-Control: no-store. FindEZ stores an Argon2 hash and cannot recover the secret. No endpoint edits an issued key's name, scopes, workspace, or expiration; replace it and revoke the old key. Do not automatically retry issuance after a timeout: it may have created a key whose secret was not received; review the key list and revoke that key before replacing it.",
        "x-required-scope": null,
        "x-rate-bucket": null,
        "x-path-values": {},
        "x-query-example": {},
        "x-extra-examples": {
          "Organization read key": {
            "request": {
              "name": "All-team reporting",
              "workspace_id": null,
              "scopes": [
                "org:read"
              ]
            }
          }
        }
      }
    },
    "/api/v1/keys/workspaces": {
      "get": {
        "tags": [
          "Key management"
        ],
        "summary": "Find teams you own",
        "operationId": "key_workspaces_api_v1_keys_workspaces_get",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OwnedWorkspaces"
                },
                "example": {
                  "workspaces": [
                    {
                      "team_id": "10000000-0000-0000-0000-000000000001",
                      "name": "Robotics team"
                    }
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Error; see the error-code reference. User-session and gateway failures may have a string detail or non-JSON body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "UserSession": []
          }
        ],
        "description": "Requires the signed-in owner's Supabase user access token. Returns teams owned by that user, ordered by name. Membership alone does not qualify. The returned team_id is the workspace_id used when creating keys. Includes empty teams. Returns an empty workspaces array if none are owned. This endpoint does not accept integration API keys.",
        "x-required-scope": null,
        "x-rate-bucket": null,
        "x-path-values": {},
        "x-query-example": {},
        "x-extra-examples": {}
      }
    },
    "/api/v1/whoami": {
      "get": {
        "tags": [
          "Connection"
        ],
        "summary": "Test authentication",
        "operationId": "whoami_api_v1_whoami_get",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Identity"
                },
                "example": {
                  "key_id": "40000000-0000-0000-0000-000000000001",
                  "workspace_id": "10000000-0000-0000-0000-000000000001",
                  "scopes": [
                    "items:read",
                    "workspace:read"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Error; see the error-code reference. User-session and gateway failures may have a string detail or non-JSON body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "IntegrationKey": []
          }
        ],
        "description": "Returns the key ID, its fixed workspace (null for an organization key), and its explicit scopes. Any valid key can call this endpoint. Success verifies authentication only: it does not prove inventory is visible, that the team is still owned, or that a write will succeed. It consumes the standard request budget.",
        "x-required-scope": null,
        "x-rate-bucket": "standard",
        "x-path-values": {},
        "x-query-example": {},
        "x-extra-examples": {}
      }
    },
    "/api/v1/keys/{key_id}": {
      "delete": {
        "tags": [
          "Key management"
        ],
        "summary": "Revoke a key",
        "operationId": "revoke_key_api_v1_keys__key_id__delete",
        "security": [
          {
            "UserSession": []
          }
        ],
        "parameters": [
          {
            "name": "key_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Key Id"
            },
            "description": "Key metadata UUID, not the raw secret or key prefix."
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevokedKey"
                },
                "example": {
                  "revoked": true,
                  "id": "40000000-0000-0000-0000-000000000001",
                  "revoked_at": "2026-09-10T12:00:00Z"
                }
              }
            }
          },
          "default": {
            "description": "Error; see the error-code reference. User-session and gateway failures may have a string detail or non-JSON body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "description": "Requires a Supabase user access token for the account that issued the key. Uses the metadata id, not the secret or prefix. Revocation is permanent; retained metadata stays in the list. The secret will be rejected on subsequent authentication. Revocation cannot undo writes already committed. Unknown, other-account, and already-revoked IDs all return 404 key_not_found; a repeated revoke does not return a second success. Create and test a replacement before revoking a key used by an active integration.",
        "x-required-scope": null,
        "x-rate-bucket": null,
        "x-path-values": {
          "key_id": "40000000-0000-0000-0000-000000000001"
        },
        "x-query-example": {},
        "x-extra-examples": {}
      }
    },
...
    "/api/v1/items/{item_id}": {
      "patch": {
        "tags": [
          "Inventory"
        ],
        "summary": "Update an item",
        "operationId": "patch_item_api_v1_items__item_id__patch",
        "security": [
          {
            "IntegrationKey": []
          }
        ],
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Item Id"
            },
            "description": "Item UUID from a create response or inventory read."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/APIItemPatch"
              },
              "example": {
                "quantity": 6,
                "notes": null
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdatedItem"
                },
                "example": {
                  "updated": true,
                  "item_id": "30000000-0000-0000-0000-000000000001"
                }
              }
            }
          },
          "default": {
            "description": "Error; see the error-code reference. User-session and gateway failures may have a string detail or non-JSON body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "description": "Updates only supplied fields. At least one field is required. name, category, quantity, and location cannot be null; nullable optional fields can be cleared with null. Quantity replaces the stored total, rather than adding a delta. Changing location moves the item to one unambiguous linked Space within the same workspace. workspace_id, space_id, user_id, item_id, and created_at are not writable here. A missing or inaccessible item returns 404. The response is an acknowledgement, not the updated record. There is no ETag, version precondition, or atomic increment: concurrent updates may overwrite one another.",
        "x-required-scope": "items:write",
        "x-rate-bucket": "standard",
        "x-path-values": {
          "item_id": "30000000-0000-0000-0000-000000000001"
        },
        "x-query-example": {},
        "x-extra-examples": {}
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "OwnedWorkspaces": {
        "type": "object",
        "properties": {
          "workspaces": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OwnedWorkspace"
            }
          }
        },
        "required": [
          "workspaces"
        ]
      },
...
      "KeyList": {
        "type": "object",
        "properties": {
          "keys": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/KeyMetadata"
            }
          }
        },
        "required": [
          "keys"
        ]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "detail": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "object",
                "properties": {
                  "code": {
                    "type": "string"
                  },
                  "message": {
                    "type": "string"
                  },
                  "correlation_id": {
                    "type": "string"
                  }
                },
                "required": [
                  "code",
                  "message",
                  "correlation_id"
                ]
              }
            ]
          }
        },
        "required": [
          "detail"
        ]
      }
    },
    "securitySchemes": {
      "IntegrationKey": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "findez_live_sk_…",
        "description": "Secret API key created in Settings → API keys. Keep server-side."
      },
      "UserSession": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Signed-in owner's Supabase access_token; only for key management. An integration API key cannot manage keys."
      }
    }
  },
  "servers": [
    {
      "url": "https://findez.openstack.ftctools.com",
      "description": "Production"
    }
  ],
...
}